Simple debug application runner
