+++ .gitignore
@@ -130,3 +130,4 @@ dmypy.json
 
 # IDEA configurations
 .idea/
+/.vscode/
