+++ docker-compose.yml
@@ -14,15 +14,15 @@ services:
     command: -p ${DB_PORT}
     network_mode: "host"
 
-  app:
-    container_name: yawm-backend-app
-    build:
-      context: .
-      dockerfile: ./docker/python/Dockerfile
-    ports:
-      - ${BACKEND_APP_PORT}:${BACKEND_APP_PORT}
-    volumes:
-      - ./backend/:${PYTHONPATH}
-    working_dir: ${PYTHONPATH}
-    command: sh -c "sleep 5 && pipenv run alembic upgrade head && pipenv run uvicorn app.main:app --reload"
-    network_mode: "host"
+#  app:
+#    container_name: yawm-backend-app
+#    build:
+#      context: .
+#      dockerfile: ./docker/python/Dockerfile
+#    ports:
+#      - ${BACKEND_APP_PORT}:${BACKEND_APP_PORT}
+#    volumes:
+#      - ./backend/:${PYTHONPATH}
+#    working_dir: ${PYTHONPATH}
+#    command: sh -c "sleep 5 && pipenv run alembic upgrade head && pipenv run uvicorn app.main:app --reload"
+#    network_mode: "host"
